Fix: Upscale TS to reduce precision loss

+++ test/recon/properties/SynchProperties.sol
@@ -20,7 +20,7 @@ abstract contract SynchProperties is BeforeAfter {
                 (uint88 votes, , uint16 epoch) = governance.lqtyAllocatedByUserToInitiative(users[j], deployedInitiatives[i]);
 
                 // Grab epoch from initiative
-                (uint88 lqtyAllocatedByUserAtEpoch, uint32 ts) =
+                (uint88 lqtyAllocatedByUserAtEpoch, uint120 ts) =
                     IBribeInitiative(deployedInitiatives[i]).lqtyAllocatedByUserAtEpoch(users[j], epoch);
 
                 // Check that TS matches (only for votes)
@@ -29,7 +29,7 @@ abstract contract SynchProperties is BeforeAfter {
                 if(votes != 0) {
                     // if we're voting and the votes are different from 0
                     // then we check user TS
-                    (, uint32 averageStakingTimestamp) = governance.userStates(users[j]);
+                    (, uint120 averageStakingTimestamp) = governance.userStates(users[j]);
 
                     eq(averageStakingTimestamp, ts, "Timestamp must be most recent when it's non zero");
                 } else {
